requirements.txt
pillow
pycryptodome
fastapi
uvicorn[standard]
python-multipart
pytest
